:truck: move `MultitenantManagerError` definition to error module
