+++ acapy_agent/multitenant/error.py
@@ -3,5 +3,9 @@
 from ..core.error import BaseError
 
 
+class MultitenantManagerError(BaseError):
+    """Generic multitenant error."""
+
+
 class WalletKeyMissingError(BaseError):
     """Wallet key missing exception."""
